moved typing check; multiple warnings for type mismatch
